chore: Update to Rust 1.97. Update Rust dependencies.

+++ src/rust/dcmfx_pixel_data/src/encode/charls.rs
@@ -217,7 +217,7 @@ fn encode(
     });
   }
 
-  output_buffer.truncate(bytes_written as usize);
+  output_buffer.truncate(bytes_written);
 
   Ok(output_buffer)
 }
